exploring_distances/pyproject.toml
[tool.poetry]
name = "exploring-distances"
version = "0.1.0"
description = ""
authors = ["Gavin Mendel-Gleason <gavin@terminusdb.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
openai = "^1.3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
